FB0T-102 liczek do osobnego modułu
